test505: int4 => int8
git-svn-id: http://data-race-test.googlecode.com/svn/trunk@589 93a3164d-3a44-0410-8ae5-4baa038b0729

+++ unittest/racecheck_unittest.cc
@@ -5586,7 +5586,7 @@ const int N_THREADS = 2,
 
 const int ARRAY_SIZE = 4 * HG_CACHE_SIZE,
           ITERATIONS = 5;
-int array[ARRAY_SIZE];
+int64_t array[ARRAY_SIZE];
 
 int count = 0;
 Mutex count_mu;
